+++ analyzer_test.go
@@ -47,6 +47,7 @@ func TestAnalyzer_Visitors(t *testing.T) {
 				Visitors:   42,
 				Sessions:   67,
 				Bounces:    30,
+				Views:      71,
 			},
 		}
 		assert.NoError(t, store.SaveVisitorStats(nil, stats))
@@ -74,6 +75,10 @@ func TestAnalyzer_Visitors(t *testing.T) {
 		assert.Equal(t, 30, visitors[1].Bounces)
 		assert.Equal(t, 0, visitors[2].Bounces)
 		assert.Equal(t, 2, visitors[3].Bounces)
+		assert.Equal(t, 0, visitors[0].Views)
+		assert.Equal(t, 71, visitors[1].Views)
+		assert.Equal(t, 0, visitors[2].Views)
+		assert.Equal(t, 2, visitors[3].Views)
 		assert.InDelta(t, 0, visitors[0].BounceRate, 0.01)
 		assert.InDelta(t, 0.71, visitors[1].BounceRate, 0.01)
 		assert.InDelta(t, 0, visitors[2].BounceRate, 0.01)
@@ -526,6 +531,7 @@ func TestAnalyzer_PageVisitors(t *testing.T) {
 				Visitors:   42,
 				Sessions:   67,
 				Bounces:    30,
+				Views:      71,
 			},
 		}
 		assert.NoError(t, store.SaveVisitorStats(nil, stats))
@@ -545,10 +551,14 @@ func TestAnalyzer_PageVisitors(t *testing.T) {
 		assert.Equal(t, 1, visitors[1].Visitors)
 		assert.Equal(t, 30, visitors[0].Bounces)
 		assert.Equal(t, 0, visitors[1].Bounces)
+		assert.Equal(t, 72, visitors[0].Views)
+		assert.Equal(t, 1, visitors[1].Views)
 		assert.InDelta(t, 0.9772, visitors[0].RelativeVisitors, 0.01)
 		assert.InDelta(t, 0.0227, visitors[1].RelativeVisitors, 0.01)
 		assert.InDelta(t, 0.6976, visitors[0].BounceRate, 0.01)
 		assert.InDelta(t, 0, visitors[1].BounceRate, 0.01)
+		assert.InDelta(t, 0.9863, visitors[0].RelativeViews, 0.01)
+		assert.InDelta(t, 0.0136, visitors[1].RelativeViews, 0.01)
 		assert.Equal(t, pastDay(3), visitors[0].Stats[0].Day)
 		assert.Equal(t, pastDay(2), visitors[0].Stats[1].Day)
 		assert.Equal(t, pastDay(1), visitors[0].Stats[2].Day)
@@ -581,6 +591,14 @@ func TestAnalyzer_PageVisitors(t *testing.T) {
 		assert.Equal(t, 0, visitors[1].Stats[1].Bounces)
 		assert.Equal(t, 0, visitors[1].Stats[2].Bounces)
 		assert.Equal(t, 0, visitors[1].Stats[3].Bounces)
+		assert.Equal(t, 0, visitors[0].Stats[0].Views)
+		assert.Equal(t, 71, visitors[0].Stats[1].Views)
+		assert.Equal(t, 0, visitors[0].Stats[2].Views)
+		assert.Equal(t, 1, visitors[0].Stats[3].Views)
+		assert.Equal(t, 0, visitors[1].Stats[0].Views)
+		assert.Equal(t, 0, visitors[1].Stats[1].Views)
+		assert.Equal(t, 0, visitors[1].Stats[2].Views)
+		assert.Equal(t, 1, visitors[1].Stats[3].Views)
 		assert.InDelta(t, 0, visitors[0].Stats[0].RelativeVisitors, 0.01)
 		assert.InDelta(t, 0.9767, visitors[0].Stats[1].RelativeVisitors, 0.01)
 		assert.InDelta(t, 0, visitors[0].Stats[2].RelativeVisitors, 0.01)
@@ -597,6 +615,14 @@ func TestAnalyzer_PageVisitors(t *testing.T) {
 		assert.InDelta(t, 0, visitors[1].Stats[1].BounceRate, 0.01)
 		assert.InDelta(t, 0, visitors[1].Stats[2].BounceRate, 0.01)
 		assert.InDelta(t, 0, visitors[1].Stats[3].BounceRate, 0.01)
+		assert.InDelta(t, 0, visitors[0].Stats[0].RelativeViews, 0.01)
+		assert.InDelta(t, 0.9861, visitors[0].Stats[1].RelativeViews, 0.01)
+		assert.InDelta(t, 0, visitors[0].Stats[2].RelativeViews, 0.01)
+		assert.InDelta(t, 0.0138, visitors[0].Stats[3].RelativeViews, 0.01)
+		assert.InDelta(t, 0, visitors[1].Stats[0].RelativeViews, 0.01)
+		assert.InDelta(t, 0, visitors[1].Stats[1].RelativeViews, 0.01)
+		assert.InDelta(t, 0, visitors[1].Stats[2].RelativeViews, 0.01)
+		assert.InDelta(t, 1, visitors[1].Stats[3].RelativeViews, 0.01)
 	}
 }
 
@@ -862,6 +888,7 @@ func TestAnalyzer_Growth(t *testing.T) {
 					Visitors: 5,
 					Sessions: 6,
 					Bounces:  3,
+					Views:    21,
 				},
 			},
 			{
@@ -871,6 +898,7 @@ func TestAnalyzer_Growth(t *testing.T) {
 					Visitors: 6,
 					Sessions: 7,
 					Bounces:  4,
+					Views:    22,
 				},
 			},
 			{
@@ -880,6 +908,7 @@ func TestAnalyzer_Growth(t *testing.T) {
 					Visitors: 2,
 					Sessions: 3,
 					Bounces:  1,
+					Views:    23,
 				},
 			},
 			{
@@ -889,6 +918,7 @@ func TestAnalyzer_Growth(t *testing.T) {
 					Visitors: 8,
 					Sessions: 9,
 					Bounces:  6,
+					Views:    24,
 				},
 			},
 		}
@@ -915,12 +945,15 @@ func TestAnalyzer_Growth(t *testing.T) {
 		assert.Equal(t, 11, growth.Current.Visitors)
 		assert.Equal(t, 13, growth.Current.Sessions)
 		assert.Equal(t, 7, growth.Current.Bounces)
+		assert.Equal(t, 43, growth.Current.Views)
 		assert.Equal(t, 10, growth.Previous.Visitors)
 		assert.Equal(t, 12, growth.Previous.Sessions)
 		assert.Equal(t, 7, growth.Previous.Bounces)
+		assert.Equal(t, 47, growth.Previous.Views)
 		assert.InDelta(t, 0.1, growth.VisitorsGrowth, 0.01)
 		assert.InDelta(t, 0.08333, growth.SessionsGrowth, 0.01)
 		assert.InDelta(t, -0.0909, growth.BouncesGrowth, 0.01)
+		assert.InDelta(t, -0.0851, growth.ViewsGrowth, 0.01)
 		growth, err = analyzer.Growth(&Filter{
 			TenantID: NewTenantID(tenantID),
 			From:     pastDay(3),
@@ -931,6 +964,7 @@ func TestAnalyzer_Growth(t *testing.T) {
 		assert.InDelta(t, 1.5, growth.VisitorsGrowth, 0.01)
 		assert.InDelta(t, 1, growth.SessionsGrowth, 0.01)
 		assert.InDelta(t, 0.1999, growth.BouncesGrowth, 0.01)
+		assert.InDelta(t, -0.0869, growth.ViewsGrowth, 0.01)
 	}
 }
 
@@ -946,12 +980,15 @@ func TestAnalyzer_GrowthNoData(t *testing.T) {
 	assert.Equal(t, 0, growth.Current.Visitors)
 	assert.Equal(t, 0, growth.Current.Sessions)
 	assert.Equal(t, 0, growth.Current.Bounces)
+	assert.Equal(t, 0, growth.Current.Views)
 	assert.Equal(t, 0, growth.Previous.Visitors)
 	assert.Equal(t, 0, growth.Previous.Sessions)
 	assert.Equal(t, 0, growth.Previous.Bounces)
+	assert.Equal(t, 0, growth.Previous.Views)
 	assert.InDelta(t, 0, growth.VisitorsGrowth, 0.01)
 	assert.InDelta(t, 0, growth.SessionsGrowth, 0.01)
 	assert.InDelta(t, 0, growth.BouncesGrowth, 0.01)
+	assert.InDelta(t, 0, growth.ViewsGrowth, 0.01)
 }
 
 func TestAnalyzer_CalculateGrowth(t *testing.T) {