+++ model.go
@@ -25,8 +25,10 @@ type Stats struct {
 	Visitors         int            `db:"visitors" json:"visitors"`
 	Sessions         int            `db:"sessions" json:"sessions"`
 	Bounces          int            `db:"bounces" json:"bounces"`
+	Views            int            `db:"views" json:"views"`
 	RelativeVisitors float64        `db:"-" json:"relative_visitors"`
 	BounceRate       float64        `db:"-" json:"bounce_rate"`
+	RelativeViews    float64        `db:"-" json:"relative_views"`
 }
 
 func (stats *Stats) visitors() int {