+++ schema/postgres/v1.13.0.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "visitor_stats" ADD COLUMN "views" integer NOT NULL DEFAULT 0;